Periodic error prone cleanup ()

+++ community/testing/wal-utils/src/main/java/org/neo4j/kernel/impl/transaction/log/InMemoryClosableChannel.java
@@ -282,7 +282,7 @@ public class InMemoryClosableChannel
     @Override
     public void position(long byteOffset) {
         var buffer = isReader ? reader : writer;
-        buffer.position((int) byteOffset);
+        buffer.position(byteOffset);
     }
 
     @Override
